% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterBoot.R
\name{clusterBoot}
\alias{clusterBoot}
\title{Function to draw cluster-level bootstrap samples}
\usage{
clusterBoot(data, n = NULL)
}
\arguments{
\item{data}{data.frame object containing cluster ID, participant ID, A1, R, A2, Y with
column names being i, j, A1, R, A2, Y.}

\item{n}{number of bootstrap samples to be drawn}
}
\value{
data.frame object with the bootstrapped samples
}
\description{
Function to draw cluster-level bootstrap samples
}
