% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered.R
\name{clustered.prepare_data}
\alias{clustered.prepare_data}
\title{Function to prepare data for estimation}
\usage{
clustered.prepare_data(data, i = "i", A1 = "A1", R = "R", A2 = "A2", Y = "Y")
}
\arguments{
\item{data}{data frame to be prepared for estimation}

\item{i}{cluster ID column name}

\item{A1}{A1 column name}

\item{R}{R column name}

\item{A2}{A2 column name}

\item{Y}{outcome variable column name}
}
\value{
a data frame ready to be plugging into the estimation function
}
\description{
Function to prepare data for estimation
}
