% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSMART.R
\name{cSMART.mm}
\alias{cSMART.mm}
\title{Function to estimate the marginal mean model parameters}
\usage{
cSMART.mm(
  formula,
  data,
  i = "i",
  A1 = "A1",
  R = "R",
  A2 = "A2",
  Y = "Y",
  covstr = "EXCH",
  verbose = T,
  weights = NULL
)
}
\arguments{
\item{formula}{formula for the estimation}

\item{data}{a data.frame object}

\item{i}{cluster ID column name}

\item{A1}{A1 column name}

\item{R}{R column name}

\item{A2}{A2 column name}

\item{Y}{outcome variable column name}

\item{covstr}{structure of the working covariance matrix 'IND' or 'EXCH' or 'EXCH|AI'}

\item{verbose}{T/F for the model fit reports}

\item{weights}{a data frame containing cluster level weights with columns `i` and `W` for cluster ID and weights respectively}
}
\value{
estimates object containing all the estimated parameters.
}
\description{
Function to estimate the marginal mean model parameters
}
