% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{estimate_betas}
\alias{estimate_betas}
\title{Function to estimate betas given the data, variance and correlation}
\usage{
estimate_betas(var, rho, data)
}
\arguments{
\item{var}{list of variances}

\item{rho}{list of correlations}

\item{data}{data with cluster ID, A1, R, A2, Y}
}
\description{
Function to estimate betas given the data, variance and correlation
}
