% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovairiates.from_conditional}
\alias{nocovairiates.from_conditional}
\title{Function to get treatment recipe from the conditionals}
\usage{
nocovairiates.from_conditional(
  cell_mu = c(10, 8, 2, 4, -2, 3),
  cell_var = c(100, 100, 100, 100, 100, 100),
  cell_cor = c(0.1, 0.2, 0.15, 0.12, 0.11, 0.07),
  N = 200,
  m = 5,
  p_1 = 0.3,
  p_2 = 0.2
)
}
\arguments{
\item{cell_mu}{a vector of conditional means}

\item{cell_var}{a vector of conditional vars}

\item{cell_cor}{a vector of conditional correlation}

\item{N}{number of clusters}

\item{m}{number of participants in a cluster}

\item{p_1}{probability of response of A_1 = 1}

\item{p_2}{probability of response of A_1 = -1}
}
\description{
Does what it says
}
