% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovariates.data_prep}
\alias{nocovariates.data_prep}
\title{Function to prepare data for estimation functions}
\usage{
nocovariates.data_prep(data, cluster_ID, participant_ID, A1, R, A2, Y)
}
\arguments{
\item{data}{data frame to be modified}

\item{cluster_ID}{the column name corresponding to the cluster ID in the data}

\item{participant_ID}{the column name corresponding to the participant ID in the data}

\item{A1}{the column name corresponding to A1 in the data}

\item{R}{the column name corresponding to R in the data}

\item{A2}{the column name corresponding to A2 in the data}

\item{Y}{the column name corresponding to the outcome variable in the data}
}
\value{
a data.frame object with renamed columns, cluster ID and participant ID.
}
\description{
nocovariates.estimate requires data as a data.frame object containing cluster ID, participant ID, A1, R, A2, Y with
column names being i, j, A1, R, A2, Y with cluster ID (i) being a contiguous integer from 1 to N and participant ID (j) being a contiguous integer from 1 to m_i
where N is the number of clusters and m_i is the number of participants in cluster i.
This helper function prepares data for estimation.
}
