% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovariates.estimate_betas}
\alias{nocovariates.estimate_betas}
\title{Function to estimate betas given the data, variance and correlation}
\usage{
nocovariates.estimate_betas(data)
}
\arguments{
\item{data}{data with cluster ID, A1, R, A2, Y}
}
\description{
Function to estimate betas given the data, variance and correlation
}
