% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovariates.estimate_effectsize}
\alias{nocovariates.estimate_effectsize}
\title{Function to estimate effect size for two DTRs}
\usage{
nocovariates.estimate_effectsize(estimates, d1, d2)
}
\arguments{
\item{estimates}{estimates object returned by the nocovariates.estimate() function}

\item{d1}{DTR 1}

\item{d2}{DTR 2}
}
\value{
estimated effect size of type numeric
}
\description{
Function to estimate effect size for two DTRs
}
