% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovariates.estimate_effectsizes}
\alias{nocovariates.estimate_effectsizes}
\title{Function to estimate effect size for all DTR pairs}
\usage{
nocovariates.estimate_effectsizes(estimates)
}
\arguments{
\item{estimates}{estimates object returned by the nocovariates.estimate() function}
}
\value{
a named list of estimated effect sizes
}
\description{
Function to estimate effect size for all DTR pairs
}
