% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovariates.estimate_var_effectsize}
\alias{nocovariates.estimate_var_effectsize}
\title{Function to estimate variance of the effect size for two DTRs}
\usage{
nocovariates.estimate_var_effectsize(estimates, d1, d2, method = "naive")
}
\arguments{
\item{estimates}{estimates object returned by the nocovariates.estimate() function}

\item{d1}{DTR 1}

\item{d2}{DTR 2}

\item{method}{can be either 'naive' or 'bootstrapped'}
}
\value{
estimated variance of the effect size of type numeric
}
\description{
Function to estimate variance of the effect size for two DTRs
}
