% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovariates.treat}
\alias{nocovariates.treat}
\title{Function to generate data from the treatment recipe object}
\usage{
nocovariates.treat(recipe)
}
\arguments{
\item{recipe}{the generated recipe from nocovairiates.from_conditional() function}
}
\value{
a data.frame object of the generated data with cluster ID, participant ID, A1, R, A2, and response
}
\description{
Generates data from the treatment recipe object. This function can be
reused with the same treatment recipe object to generate more samples with the same parameters.
}
