% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovariates.true_effect_size}
\alias{nocovariates.true_effect_size}
\title{Function to output the true effect size from the recipe}
\usage{
nocovariates.true_effect_size(recipe, d1, d2)
}
\arguments{
\item{recipe}{the generated recipe from nocovairiates.from_conditional() function}

\item{d1}{DTR 1}

\item{d2}{DTR 2}
}
\value{
the true effect size of two DTRs
}
\description{
Function to output the true effect size from the recipe
}
