% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_summary.R
\name{treatment_summary}
\alias{treatment_summary}
\title{Function to get treatment summary from the data generator}
\usage{
treatment_summary(recipe)
}
\arguments{
\item{recipe}{the generated recipe from nocovairiates.from_conditional()}
}
\description{
Function to get treatment summary from the data generator
}
