% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovariates.estimate}
\alias{nocovariates.estimate}
\title{Function to estimate beta, variance of beta_hat, and ICC}
\usage{
nocovariates.estimate(data)
}
\arguments{
\item{data}{a data.frame object containing cluster ID, participant ID, A1, R, A2, Y with
column names being i, j, A1, R, A2, Y.}
}
\value{
a list object containing all the estimated parameters.
}
\description{
Function to estimate beta, variance of beta_hat, and ICC
}
