% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovariates.estimate_var_effectsize}
\alias{nocovariates.estimate_var_effectsize}
\title{Function to estimate variance of the effect size for two DTRs}
\usage{
nocovariates.estimate_var_effectsize(
  estimates,
  d1,
  d2,
  data,
  method = "naive",
  numBoot = 1000,
  numCores = 1
)
}
\arguments{
\item{estimates}{estimates object returned by the nocovariates.estimate() function}

\item{d1}{DTR 1}

\item{d2}{DTR 2}

\item{data}{data to be passed as an argument for 'bootstrap' method}

\item{method}{can be either 'naive' or 'bootstrap'}

\item{numBoot}{the number of bootstrap simulations to be performed}

\item{numCores}{number of parallel cores to be passed for parallel processing}
}
\value{
estimated variance of the effect size of type numeric
}
\description{
Function to estimate variance of the effect size for two DTRs
}
