% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocovariates.R
\name{nocovariates.treat}
\alias{nocovariates.treat}
\title{Function to generate data from treatment}
\usage{
nocovariates.treat(recipe)
}
\arguments{
\item{recipe}{the generated recipe from nocovairiates.from_conditional()}
}
\description{
Function to generate data from treatment
}
